Address feedback
- Use a string builder instead of string to minimize allocations
- Use proper doc-comment format

With the use of stringBuilder, errors have been introduced into the package.
I added the dependency `github.com/pkg/errors` to help handle them, but if
we'd prefer to stick with the standard library I think that'd be fair. My
thinking was the circumstances that caused strings.Builder.WriteString() to
error would warrent a stack trace and metadata, something which is more
pragmatically done using `github.com/pkg/errors`.